Move filter logic to its own package
