+++ README.md
@@ -37,6 +37,8 @@ This API key is not provided here due to liability concerns, but it appears to b
 
 ## Example Usage
 
+Example can be seen in [`example/main.go`](example/main.go) or below:
+
 ```go
 package main
 
@@ -47,6 +49,7 @@ import (
 	"time"
 
 	"github.com/ahobsonsayers/twigots"
+	"github.com/ahobsonsayers/twigots/filter"
 )
 
 func main() {
@@ -73,9 +76,9 @@ func main() {
 	log.Printf("Fetched %d ticket listings", len(listings))
 
 	// Filter ticket listings just by name
-	hamiltonListings := twigots.FilterTicketListings(
+	hamiltonListings := filter.FilterTicketListings(
 		listings,
-		twigots.EventNamePredicate("Hamilton", 0.9), // Similarity of 0.9 - allow minor mismatches
+		filter.EventName("Hamilton", 0.9), // Similarity of 0.9 - allow minor mismatches
 	)
 	for _, listing := range hamiltonListings {
 		slog.Info(
@@ -89,15 +92,15 @@ func main() {
 	}
 
 	// Filter ticket listings just by several filters
-	coldplayListings := twigots.FilterTicketListings(
+	coldplayListings := filter.FilterTicketListings(
 		listings,
-		twigots.EventNamePredicate("Coldplay", 1), // Similarity of 1 - exact match only
-		twigots.RegionsPredicate( // Only in specific regions
+		filter.EventName("Coldplay", 1), // Similarity of 1 - exact match only
+		filter.EventRegion( // Only in specific regions
 			twigots.RegionLondon,
 			twigots.RegionSouth,
 		),
-		twigots.NumTicketsPredicate(2),    // Exactly 2 tickets in the listing
-		twigots.MinDiscountPredicate(0.1), // Discount of > 10%
+		filter.NumTickets(2),    // Exactly 2 tickets in the listing
+		filter.MinDiscount(0.1), // Discount of > 10%
 	)
 	for _, listing := range coldplayListings {
 		slog.Info(
