+++ client.go
@@ -185,11 +185,9 @@ func sliceToMaxNumTicketListings(listings TicketListings, maxNumTicketListings i
 }
 
 func filterToCreatedAfter(listings TicketListings, createdAfter time.Time) TicketListings {
-	createdAfterPredicate := CreatedAfterPredicate(createdAfter)
 	filteredListings := make(TicketListings, 0, len(listings))
 	for _, listing := range listings {
-		ok := createdAfterPredicate(listing)
-		if ok {
+		if listing.CreatedAt.Time.After(createdAfter) {
 			filteredListings = append(filteredListings, listing)
 		}
 	}
